Cambiado parámetro celular de tipo Double a tipo Long para evitar problemas

+++ src/main/java/com/alnicode/platzi/market/persistence/entity/Cliente.java
@@ -22,7 +22,7 @@ public class Cliente {
 
     private String apellidos;
 
-    private Double celular;
+    private Long celular;
 
     private String direccion;
